# src/main/resources/static/docs/trycatch_openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "trycatch API",
    "description" : "trycatch API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://trycatch"
  } ],
  "paths" : {
...
    "/<unknown>/question" : {
      "get" : {
        "summary" : "GET /v1/question",
        "operationId" : "findAllQuestions",
        "parameters" : [ {
          "name" : "category",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "POST /v1/question",
        "operationId" : "createQuestion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateQuestionRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/<unknown>/question/{questionId}" : {
      "get" : {
        "summary" : "GET /v1/question/{questionId}",
        "operationId" : "findQuestionById",
        "parameters" : [ {
          "name" : "questionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      },
      "put" : {
        "summary" : "PUT /v1/question/{questionId}",
        "operationId" : "putQuestion",
        "parameters" : [ {
          "name" : "questionId",
          "in" : "path",
          "required" : false
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PutQuestionRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "DELETE /v1/question/{questionId}",
        "operationId" : "deleteQuestion",
        "parameters" : [ {
          "name" : "questionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/question/{questionId}/answer" : {
      "put" : {
        "summary" : "PUT /v1/question/{questionId}/answer",
        "operationId" : "putAnswer",
        "parameters" : [ {
          "name" : "questionId",
          "in" : "path",
          "required" : false
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PutAnswerRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "POST /v1/question/{questionId}/answer",
        "operationId" : "createAnswers",
        "parameters" : [ {
          "name" : "questionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAnswerRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/question/{questionId}/{answerId}" : {
      "post" : {
        "summary" : "POST /v1/question/{questionId}/{answerId}",
        "operationId" : "acceptAnswer",
        "parameters" : [ {
          "name" : "questionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "answerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "HttpHeaders" : {
        "type" : "object",
        "properties" : { }
      },
      "ResponseEntity" : {
        "type" : "object",
        "properties" : {
          "headers" : {
            "$ref" : "#/components/schemas/HttpHeaders"
          }
        }
      },
      "CreateQuestionRequestDto" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "PutQuestionRequestDto" : {
        "type" : "object",
        "properties" : {
          "questionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "category" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "hidden" : {
            "type" : "boolean"
          }
        }
      },
      "PutAnswerRequestDto" : {
        "type" : "object",
        "properties" : {
          "answerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "content" : {
            "type" : "string"
          },
          "hidden" : {
            "type" : "boolean"
          }
        }
      },
      "CreateAnswerRequestDto" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}